First major commit!!
Plugin, Sub-Plugin, and Package implementations for list, create, and view all work.

Zip file validation currently only checks for the primary directory and needs a lot of updating.

Side-bar is working, but needs 'prettied-up' and extended with more functionality.

None of the Admin implementation is written, and will remain that way until testing on a live server with phpBB3 is available.  Same goes for extending the User functionality.